Modified as per discussion
