+++ lib/philomena_web/router.ex
@@ -119,7 +119,13 @@ defmodule PhilomenaWeb.Router do
     resources "/comments", CommentController, only: [:show]
     resources "/posts", PostController, only: [:show]
     resources "/profiles", ProfileController, only: [:show]
-    resources "/filters", FilterController, only: [:show, :index]
+
+    scope "/filters", Filter, as: :filter do
+      resources "/user", UserFilterController, only: [:index]
+      resources "/system", SystemFilterController, only: [:index]
+    end
+
+    resources "/filters", FilterController, only: [:show]
 
     resources "/forums", ForumController, only: [:show, :index] do
       resources "/topics", Forum.TopicController, only: [:show, :index] do
